+++ server/items/bullets/banana_bullet.cpp
@@ -9,7 +9,7 @@
 #include <iostream>
 
 BananaBullet::BananaBullet(uint8_t type, uint8_t id, float x_pos, float y_pos,
-                           int damage, uint8_t range, float spread)
+                           int damage, int range, float spread)
     : Bullet(type, id, x_pos, y_pos, damage, range, spread),
       velocidad(VELOCIDAD_INICIAL), continue_moving(true) {}
