+++ server/items/weapons/duel_pistol.cpp
@@ -1,7 +1,7 @@
 #include "duel_pistol.h"
 
 DuelPistol::DuelPistol(uint8_t type, uint8_t id, float x_pos, float y_pos,
-                       int damage, uint8_t range, int ammo_quantity,
+                       int damage, int range, int ammo_quantity,
                        float recoil)
     : Weapon(type, id, x_pos, y_pos, damage, range, ammo_quantity, recoil),
       bala(PISTOLA_DUELOS_BULLET, 1, 0, 0, damage, range, 0.3) {}
