+++ server/items/weapons/granada.h
@@ -16,7 +16,7 @@ private:
 
 public:
   Granada(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
-          uint8_t range, int ammo_quantity, float recoil);
+          int range, int ammo_quantity, float recoil);
   bool isEmptyAmmo() override;
   std::unique_ptr<Bullet> shoot(bool is_aiming_up) override;
   bool isActive() override;
