+++ server/items/weapons/weapon.h
@@ -10,7 +10,7 @@
 class Weapon : public Objeto {
 protected:
   int damage;
-  uint8_t range;
+  int range;
   int ammo_quantity;
   float recoil;
   int reload_time;
@@ -18,7 +18,7 @@ protected:
 
 public:
   Weapon(uint8_t type, uint8_t id, float x_pos, float y_pos, int damage,
-         uint8_t range, int ammo_quantity, float recoil);
+         int range, int ammo_quantity, float recoil);
 
   virtual bool isEmptyAmmo() = 0;
   virtual std::unique_ptr<Bullet> shoot(bool is_aiming_up) = 0;
